chore(coinjoin): move enums to standalone file
