+++ packages/coinjoin/src/enums.ts
@@ -1,0 +1,17 @@
+export enum RoundPhase {
+    InputRegistration = 0,
+    ConnectionConfirmation = 1,
+    OutputRegistration = 2,
+    TransactionSigning = 3,
+    Ended = 4,
+}
+
+export enum EndRoundState {
+    None = 0,
+    AbortedWithError = 1,
+    AbortedNotEnoughAlices = 2,
+    TransactionBroadcastFailed = 3,
+    TransactionBroadcasted = 4,
+    NotAllAlicesSign = 5,
+    AbortedNotEnoughAlicesSigned = 6,
+}
