+++ packages/coinjoin/src/types/coordinator.ts
@@ -1,3 +1,5 @@
+import { RoundPhase, EndRoundState } from '../enums';
+
 export interface CoinjoinStatus {
     roundStates: Round[];
     coinJoinFeeRateMedians: FeeRateMedians[];
@@ -127,24 +129,6 @@ export interface CoinjoinState {
     events: CoinjoinStateEvent[];
     isFullySigned?: boolean;
     witnesses?: Record<number, string>;
-}
-
-export enum RoundPhase {
-    InputRegistration = 0,
-    ConnectionConfirmation = 1,
-    OutputRegistration = 2,
-    TransactionSigning = 3,
-    Ended = 4,
-}
-
-export enum EndRoundState {
-    None = 0,
-    AbortedWithError = 1,
-    AbortedNotEnoughAlices = 2,
-    TransactionBroadcastFailed = 3,
-    TransactionBroadcasted = 4,
-    NotAllAlicesSign = 5,
-    AbortedNotEnoughAlicesSigned = 6,
 }
 
 export interface FeeRateMedians {
